% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFuelHeat.R
\name{analyzeFuelHeat}
\alias{analyzeFuelHeat}
\title{analyzeFuelHeat}
\usage{
analyzeFuelHeat(
  plantList,
  data,
  dom_thrsh,
  maj_thrsh,
  fuel.codes,
  sheet3_key,
  sheet4_key,
  combogencoo_key,
  combogencoo_cooling_type_df,
  page3 = FALSE
)
}
\arguments{
\item{plantList}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{data}{data.frame with monthly Quantity.of.Fuel.Consumed and MMBTu_per_unit, names may vary}

\item{dom_thrsh}{numeric threshold for dominant fuels}

\item{maj_thrsh}{numeric threshold for boiler majority fuel}

\item{fuel.codes}{Fuel generalization list defined as
\code{list(biomass = c("AB",'BLQ','MSB','MSN','OBL','OBS','SLW','TDF','WDS'), coal = c("ANT","BIT","LIG","PC","SUB","WC"), gas = c("BFG","LFG","NG","OBG","OG","PG","SGC","SGP"), oil = c("DFO","JF","KER","RFO","WO"), other = c("OTH"))}}

\item{page3}{logical indicating whether page 3 boilerFuelData is used as input, if FALSE page 1 gen_fuel_data is used}
}
\value{
\code{raw_fuel_heat_df_2} annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero consmption,
assigns dominant fuel codes
}
\description{
Calculates annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero consmption,
assigns dominant fuel codes.  Returns object for output files fuel_heat_table.csv and
fuel_heat_ST_only.csv. \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
