% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNominalLosses.R
\name{analyzeNominalLosses}
\alias{analyzeNominalLosses}
\title{analyzeNominalLosses}
\usage{
analyzeNominalLosses(fuel_heat_df, page3)
}
\arguments{
\item{fuel_heat_df}{annual fuel-specific fuel heat, and boiler total fuel heat at each boiler, flags
dominant fuels, majority of boiler fuel, and fuels with zero consmption, assigns
dominant fuel codes, output of the analyzeFuelHeat() function}

\item{page3}{logical indicating whether page 3 boilerFuelData is used as input, if FALSE page 1 gen_fuel_data is used}
}
\value{
\code{nominal_losses} data.frame monthly summed nominal losses by plant and boiler
}
\description{
Calculate monthly summed nominal losses by plant and boiler, for output file nominal_losses.csv \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
