% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCoolTypes.R
\name{applyCoolTypes}
\alias{applyCoolTypes}
\title{applyCoolTypes}
\usage{
applyCoolTypes(coolType_df, cd_df, coolTypes, colNames, keepCols, joinCol)
}
\arguments{
\item{coolType_df}{data.frame with cooling types and percent allocations by combogencoo}

\item{cd_df}{data.frame with the condenser duty values}

\item{coolTypes}{character vector of cooling types to apply}

\item{colNames}{character vector of column names from \code{coolType_df} that correspond to \code{coolTypes}}

\item{keepCols}{character vector of column names from the \code{coolType_df} to keep}

\item{joinCol}{character string indicating column to join \code{coolType_df} and \code{cd_df}}
}
\value{
\code{all_final_duties} data.frame that combines condenser dutes for all \code{coolTypes}
}
\description{
applies percent allocations of cooling types to condenser duty values.  For
complex cooling plants both percentage and 100 percent values are given \cr \cr
Executed By: assignCooling.R \cr
}
