% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCoolTypes_plant.R
\name{applyCoolTypes_plant}
\alias{applyCoolTypes_plant}
\title{applyCoolTypes_plant}
\usage{
applyCoolTypes_plant(
  coolType_df,
  plantData,
  coolTypes,
  colNames,
  keepCols_cool,
  keepCols_data,
  joinCol,
  sumCols = NA
)
}
\description{
applyCoolTypes_plant
}
