% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignBogen.R
\name{assignBogen}
\alias{assignBogen}
\title{assignBogen}
\usage{
assignBogen(
  nominal_losses,
  nominal_losses_page1,
  netgen,
  netgen_page1,
  fuel_heat_table,
  fuel_heat_table_page1,
  final_steam_heat,
  final_steam_heat_page1,
  sheet3_key,
  sheet4_key
)
}
\arguments{
\item{nominal_losses}{Page3 monthly summed nominal losses by plant and boiler, output from the
analyzeNominalLosses() function}

\item{nominal_losses_page1}{Page1 monthly summed nominal losses by plant and Reported.Prime.Mover,
output from the analyzeNominalLosses() function}

\item{netgen}{Page4 net generation data with negatives replaced with zeros and generators
with negative annual net generation removed}

\item{netgen_page1}{Page1 net generation data with negatives replaced with zeros and generators
with negative annual net generation removed}

\item{fuel_heat_table}{Page3 annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero
consmption, assigns dominant fuel codes, output of the analyzeFuelHeat()
function}

\item{fuel_heat_table_page1}{Page1 annual fuel-specific fuel heat, and RPM total fuel heat at each RPM,
flags dominant fuels, majority of RPM fuel, and fuels with zero
consmption, assigns dominant fuel codes, output of the analyzeFuelHeat()
function}

\item{final_steam_heat}{Page3 monthly steam heat by plant and boiler, SteamHeatanalysis() function output}

\item{final_steam_heat_page1}{Page1 monthly steam heat by plant and RPM, SteamHeatanalysis() function output}

\item{sheet3_key}{Bogen associations by Boiler, executeBogenAssoc() function output}

\item{sheet4_key}{Bogen associations by Generator, executeBogenAssoc() function output}
}
\value{
\code{results} list containing bogen_fuel_heat, bogen_nom_loss, bogen_netgen, bogen_steam_heat
}
\description{
calculates monthly fuel heat by bogen, assigns bogen values to the variable
data frames used in the condenser duty calculation. \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
