% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCooling.R
\name{assignCooling}
\alias{assignCooling}
\title{assignCooling}
\usage{
assignCooling(
  sheet4_key,
  bogencoo.key,
  combogencoo_cooly_type,
  nuke_combogencoo_key,
  non_nuke_CD_summary,
  nuke_CD_summary,
  outputData_path
)
}
\arguments{
\item{sheet4_key}{plant_gen to bogen associations}

\item{bogencoo.key}{associates boilers-generators-cooling systems}

\item{combogencoo_cooly_type}{List of bogencoo's cooling types for 2015 and 2010}

\item{nuke_combogencoo_key}{nuclear plant associate boilers-generators-cooling systems}

\item{non_nuke_CD_summary}{condenser duty model execution for non nuclear plants, output
of the non_nuke_condenser_duty() function}

\item{nuke_CD_summary}{condenser duty model execution for nuclear plants, output of the
NukeCDcal() function}

\item{outputData_path}{character string indicating path for output files}
}
\value{
\code{results} data.frame that combines Steam Condenser Duties, NGCC Condenser Duties, Nuclear Conenser Duties,
and Cooling System Data
}
\description{
combines Steam Condenser Duties, NGCC Condenser Duties, Nuclear Conenser Duties,
and Cooling System Data. \cr \cr
Outputs the following files: \itemize{\item non_nuke_duties_w_cooling.csv
\item nuke_duties_w_cooling.csv
\item final_CD_w_nukes_cooling.csv
\item no_dupe_final_CD_w_nukes_cooling.csv} \cr
Executed By: CondenserDutyModel_new.R \cr
}
