% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakStr.R
\name{breakStr}
\alias{breakStr}
\title{breakStr}
\usage{
breakStr(longStr, nBreak, strBreak)
}
\arguments{
\item{longStr}{long character string}

\item{nBreak}{number of characters in between breaks}

\item{strBreak}{character string to place at break (i.e. \code{"\\n"} to enter line break)}
}
\value{
\code{newstr} character string with breaks added
}
\description{
Breaks long character strings at \code{nBreak} number of characters
with the character string indicated by \code{strbreak} \\cr \\cr
}
