% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNominalLoss.R
\name{calNominalLoss}
\alias{calNominalLoss}
\title{calNominalLoss}
\usage{
calNominalLoss(prime_mover, fuel_heat)
}
\arguments{
\item{prime_mover}{Reported.Prime.Mover from fuel_heat_table.csv}

\item{fuel_heat}{annual fuel-specific fuel heat, and boiler total fuel heat at each boiler, flags
dominant fuels, majority of boiler fuel, and fuels with zero consmption, assigns
dominant fuel codes, output of the analyzeFuelHeat() function}
}
\value{
\code{nominal_loss} data.frame containing monthly nominal losses
}
\description{
calculate nominal losses \cr \cr
Executed By: analyzeNominalLosses.R \cr
}
