% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePostProcessVars.R
\name{combinePostProcessVars}
\alias{combinePostProcessVars}
\title{combinePostProcessVars}
\usage{
combinePostProcessVars(
  plantDom,
  plantDom_page1,
  bogen_netgen,
  bogen_fuel_heat,
  gen_fuel_data,
  combogencoo_key,
  combogencoo_cooly_type,
  combogencoo_cooly_type_nukes,
  sheet4_key,
  nuke_combogencoo_key,
  plantInfo,
  plantList
)
}
\arguments{
\item{plantList}{}
}
\description{
combinePostProcessVars
}
