% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileCDintermediate.R
\name{compileCDintermediate}
\alias{compileCDintermediate}
\title{compileCDintermediate}
\usage{
compileCDintermediate(
  bogen_fuel_heat,
  bogen_netgen,
  bogen_nom_loss,
  bogen_steam_heat,
  gen_fuel_data,
  non_nuke_CD_summary,
  nuke_CD_summary,
  plantList,
  plant_exports
)
}
\arguments{
\item{plant_exports}{}
}
\description{
compileCDintermediate
}
