% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFEWSRoutput.R
\name{compileFEWSRoutput}
\alias{compileFEWSRoutput}
\title{compileFEWSRoutput}
\usage{
compileFEWSRoutput(
  FEWSRinputData_path,
  compileOutFEWSRpath,
  eia_year,
  types = c("Big_Lake", "Pond", "River", "Saline")
)
}
\arguments{
\item{types}{}
}
\description{
compileFEWSRoutput
}
