% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIAcomparePlots.R
\name{EIAcomparePlots}
\alias{EIAcomparePlots}
\title{EIAcomparePlots}
\usage{
EIAcomparePlots(FEWSRtower.list, list.8D, outputData_pathYear)
}
\arguments{
\item{outputData_pathYear}{}
}
\description{
EIAcomparePlots
}
