% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estEfficiency.R
\name{estEfficiency}
\alias{estEfficiency}
\title{estEfficiency}
\usage{
estEfficiency(ST_fuel_heat_df, boilerDesignData, page3)
}
\arguments{
\item{ST_fuel_heat_df}{data.frame subset of fuel_heat_table.csv, with only
Reported.Prime.Mover=="ST"}

\item{boilerDesignData}{data.frame 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters', input file 2015_BoilerDesignInfo.csv}

\item{page3}{logical indicating whether page 3 boilerFuelData is used as input, if FALSE page 1 gen_fuel_data is used}
}
\value{
\code{results} list of estimated boiler efficiencies and reported steam heat efficiencies
}
\description{
get reported boiler efficiencies and estimate boiler efficiencies \cr \cr
Executed By: steamHeatAnalysis.R \cr
}
