% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedDomFuel.R
\name{groupedDomFuel}
\alias{groupedDomFuel}
\title{groupedDomFuel}
\usage{
groupedDomFuel(
  df,
  groupCols,
  domGroup,
  naString = "other",
  dom_thrsh,
  maj_thrsh,
  nameTotal,
  fuel.codes
)
}
\arguments{
\item{fuel.codes}{}
}
\description{
groupedDomFuel
}
