% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NukeCDcal.R
\name{NukeCDcal}
\alias{NukeCDcal}
\title{NukeCDcal}
\usage{
NukeCDcal(gen_fuel_data)
}
\arguments{
\item{gen_fuel_data}{data.frame from EIA-923 Monthly Generation and Fuel Consumption Time Series
File, 2015 Final Revision}
}
\value{
\code{cd_results} data.frame of monthly condenser duty values for nuclear plants by plant and
nuclear unit id
}
\description{
condenser duty model execution for nuclear plants \cr \cr
Executed By: CondenserDutyModel_new.R \cr
}
