% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCoolType.R
\name{prepCoolType}
\alias{prepCoolType}
\title{prepCoolType}
\usage{
prepCoolType(sheet4_key, combogencoo_key, combogencoo_cooling_type_df)
}
\arguments{
\item{sheet4_key}{plant_gen to bogen associations}

\item{combogencoo_key}{associates boilers-generators-cooling systems}

\item{combogencoo_cooling_type_df}{List of bogencoo's cooling types for 2015 and 2010
combogencoo_paired_cooling<-prepCoolType(sheet4_key, combogencoo_key, combogencoo_cooling_type_df)}
}
\description{
reformat cooling type and join with bogen and combogencoo key\cr \cr
Executed By: \itemize{\item processExports.R
\item assignCooling.R} \cr
}
