% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExports.R
\name{processExports}
\alias{processExports}
\title{processExports}
\usage{
processExports(
  plantList,
  plant_exports,
  non_nuke_cd,
  sheet4_key,
  bogencoo.key,
  combogencoo_cooly_type,
  outputData_path
)
}
\arguments{
\item{plantList}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{plant_exports}{monthly heat exports by plant, output of the calHeatExports() function
in the heatExport_fun.R}

\item{non_nuke_cd}{combined steam and NGCC condenser duty output of analyzeSteamCD() and analyzeNGCCCD()}

\item{sheet4_key}{plant_gen to bogen associations}

\item{bogencoo.key}{associates boilers-generators-cooling systems}

\item{combogencoo_cooly_type}{List of bogencoo's cooling types for 2015 and 2010}

\item{outputData_path}{character string indicating path for output files}
}
\value{
\code{non_nuke_cd_exports} data.frame containing monthly net condenser duty heat exports for
non-nuclear plants, output of the processExports() function
}
\description{
calculates monthly net condenser duty heat exports for non-nuclear plants \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
