% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFormatTowerFEWSRresults.R
\name{readFormatTowerFEWSRresults}
\alias{readFormatTowerFEWSRresults}
\title{readFormatTowerFEWSRresults}
\usage{
readFormatTowerFEWSRresults(
  TowerinputData_path,
  FEWSRinputData_path,
  eia_year,
  types = c("Big_Lake", "Pond", "River", "Saline")
)
}
\arguments{
\item{types}{}
}
\description{
readFormatTowerFEWSRresults
}
