% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCondenserDuty.R
\name{runCondenserDuty}
\alias{runCondenserDuty}
\title{runCondenserDuty}
\usage{
runCondenserDuty(outputData_path, inputData.list, bogen.out.list)
}
\arguments{
\item{outputData_path}{character string indicating path for output files}
}
\description{
This model calculates condenser duty at nuclear and non-nuclear plants modeled
for water use estimation. Condenser duty calculations are based on data from the
2015 editions of the EIA860 and EIA923 databases on net power generation and fuel
consumption at power plants, as well as associations for boiler, generators,and
cooling systems. Citations for the EIA reports are included below. \cr \cr
Executed By: ThermoElectricWaterConsumptionModels.R \cr
}
