% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalEfficiencyByBogen.R
\name{thermalEfficiencyByBogen}
\alias{thermalEfficiencyByBogen}
\title{thermalEfficiencyByBogen}
\usage{
thermalEfficiencyByBogen(bogen_fuel_heat, bogen_netgen, gen_fuel_data)
}
\description{
thermalEfficiencyByBogen
}
