% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackList.R
\name{unPackList}
\alias{unPackList}
\title{unPackList}
\usage{
unPackList(lists, parentObj, env = parent.frame())
}
\arguments{
\item{lists}{list of variables to unpack into the \code{parent.frame()}}

\item{parentObj}{list of parent objects that \code{lists} are part of, if NA no parent object}
}
\description{
creates individual variables from list objects and places them in the
\code{parent.frame()} environment \\cr \\cr
}
